Adapt longest movie title tests - use int instead of long as expected result

+++ src/main/test/at/ac/fhcampuswien/fhmdb/GetLongestMovieTitleTest.java
@@ -14,13 +14,13 @@ public class GetLongestMovieTitleTest extends BaseTest {
     @Test
     void longest_movie_title_with_all_movies_given_should_return_46() {
         // Star Wars: Episode V - The Empire Strikes Back
-        long result = MovieUtils.getLongestMovieTitle(movieList);
+        int result = MovieUtils.getLongestMovieTitle(movieList);
         assertEquals(46, result);
     }
 
     @Test
     void longest_movie_title_with_list_of_movies_containing_The_Godfather_and_Pulp_Fiction_should_return_13() {
-        long result = MovieUtils.getLongestMovieTitle(
+        int result = MovieUtils.getLongestMovieTitle(
                 List.of(movieMap.get("Pulp Fiction"), movieMap.get("The Godfather"))
         );
         assertEquals(13, result);
@@ -28,7 +28,7 @@ public class GetLongestMovieTitleTest extends BaseTest {
 
     @Test
     void longest_movie_title_with_list_of_movies_containing_The_Shawshank_Redemption_Once_Upon_a_Time_in_Hollywood_12_Angry_Men_should_return_29() {
-        long result = MovieUtils.getLongestMovieTitle(
+        int result = MovieUtils.getLongestMovieTitle(
                 List.of(movieMap.get("The Shawshank Redemption"),
                         movieMap.get("Once Upon a Time in Hollywood"),
                         movieMap.get("12 Angry Men"))
@@ -41,19 +41,19 @@ public class GetLongestMovieTitleTest extends BaseTest {
         Movie noTitleMovie = new Movie.Builder()
                 .setTitle("")
                 .build();
-        long result = MovieUtils.getLongestMovieTitle(List.of(noTitleMovie));
+        int result = MovieUtils.getLongestMovieTitle(List.of(noTitleMovie));
         assertEquals(0, result);
     }
 
     @Test
     void longest_movie_title_with_empty_list_should_return_0() {
-        long result = MovieUtils.getLongestMovieTitle(new ArrayList<>());
+        int result = MovieUtils.getLongestMovieTitle(new ArrayList<>());
         assertEquals(0, result);
     }
 
     @Test
     void longest_movie_title_with_null_list_should_return_0() {
-        long result = MovieUtils.getLongestMovieTitle(null);
+        int result = MovieUtils.getLongestMovieTitle(null);
         assertEquals(0, result);
     }
